services/crop-advisor/requirements.txt
langchain>=0.1.0
langchain-groq>=0.0.1
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
